refactor: removed misplaced classes from domain
