+++ src/main/kotlin/carRace/domain/RaceIterationProvider.kt
@@ -3,13 +3,7 @@ package carRace.domain
 interface RaceIterationProvider {
     fun provide(): Int
 
-    class Manual(private val iteration: Int) : RaceIterationProvider {
+    class ManualRaceIterationProvider(private val iteration: Int) : RaceIterationProvider {
         override fun provide(): Int = iteration
-    }
-
-    object UserInput : RaceIterationProvider {
-        override fun provide(): Int {
-            return println("시도할 횟수는 몇 회인가요?").run { readln().trim().toInt() }
-        }
     }
 }
